Setting to run in COBA not CUBA
From what I understand, the tsodyks-markram synapse model that the CA3 example uses requires the simulation to be run in conductance mode. Also, even with code modifications, setting the mode to COBA instead of CUBA is needed to allow for synapse receptor-specific signal processing that is needed in the model. I am not familiar with the code enough to say for sure but for now I suggest trying this mode by default.